+++ src/nodes/args.js
@@ -3,6 +3,7 @@ const {
   group,
   ifBreak,
   indent,
+  literalline,
   join,
   line,
   softline
@@ -11,23 +12,94 @@ const {
 const toProc = require("../toProc");
 const { makeArgs } = require("../utils");
 
+// This handles a stupidly specific case where you have heredocs as arguments in
+// addition to have a block being passed. In that case we can't do anything
+// fancy with the heredoc beginnings because we need to have the braces or
+// do..end available for the block. The code that would hit this path looks like
+// this:
+//
+//     my_method(<<~HEREDOC
+//       Hello
+//     HEREDOC
+//     ) { "stuff passed to block" }
+//
+// I don't know why you would do this instead of having the heredoc split out
+// into a variable, but since this was reported as an issue (#301) I'm going to
+// explicitly handle it.
+function handleHeredocAndBlockInArgParen(path, { addTrailingCommas }, print) {
+  let argsNode = path.getValue().body[0];
+  let argsPath = [print, "body", 0];
+
+  if (argsNode.type === "args_add_block") {
+    argsNode = argsNode.body[0];
+    argsPath.push("body", 0);
+  }
+
+  // If there aren't any heredocs to be found, then we can just return null.
+  if (argsNode.body.every((node) => node.type !== "heredoc")) {
+    return null;
+  }
+
+  // If 2 nodes up in the tree we have a method_add_block node, then we know we
+  // are passing a block and need to have special handling. Otherwise we can
+  // just return null.
+  if (path.getParentNode(1).type !== "method_add_block") {
+    return null;
+  }
+
+  const parts = [];
+  const argsLength = argsNode.body.length;
+
+  argsNode.body.forEach((argNode, index) => {
+    const isInner = index !== argsLength - 1;
+
+    if (argNode.type === "heredoc") {
+      parts.push(
+        argNode.beging,
+        isInner || addTrailingCommas ? "," : "",
+        literalline,
+        concat(path.map.apply(path, argsPath.concat("body", index, "body"))),
+        argNode.ending,
+        isInner ? line : ""
+      );
+    } else {
+      parts.push(path.call.apply(path, argsPath.concat("body", index)));
+
+      if (isInner) {
+        parts.push(concat([",", line]));
+      } else if (addTrailingCommas) {
+        parts.push(",");
+      }
+    }
+  });
+
+  return group(concat(["(", indent(concat(parts)), softline, ")"]));
+}
+
 module.exports = {
   arg_paren: (path, opts, print) => {
-    if (path.getValue().body[0] === null) {
+    const argsNode = path.getValue().body[0];
+    const { addTrailingCommas } = opts;
+
+    if (argsNode === null) {
       return "";
     }
 
     // Here we can skip the entire rest of the method by just checking if it's
     // an args_forward node, as we're guaranteed that there are no other arg
     // nodes.
-    if (path.getValue().body[0].type === "args_forward") {
+    if (argsNode.type === "args_forward") {
       return "(...)";
     }
 
-    const { addTrailingCommas } = opts;
-    const { args, heredocs } = makeArgs(path, opts, print, 0);
+    // If we hit a very special case where we have heredocs and are surrounded
+    // by a method call that passes a block, then we handle that here.
+    const printed = handleHeredocAndBlockInArgParen(path, opts, print);
+    if (printed) {
+      return printed;
+    }
 
-    const argsNode = path.getValue().body[0];
+    const { args, heredocs } = makeArgs(path, opts, print, 0);
     const hasBlock = argsNode.type === "args_add_block" && argsNode.body[1];
 
     if (heredocs.length > 1) {